pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sbm-harness"
version = "0.1.0"
description = "Safe Bounded Memory (SBM) Test Harness with Physics-Based Fault Injection"
requires-python = ">=3.11"
dependencies = [
    "PyYAML==6.0.1",
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-snapshot==0.9.0",
    "coverage==7.3.2",
    "jsonschema==4.20.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=.",
    "--cov-branch",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
branch = true
source = ["."]
omit = [
    "tests/*",
    "build/*",
    "*/site-packages/*",
    ".venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "build/coverage_html"
